Move test require cgroups checks to require_linux.go
New golangci-lint errors instead of warns on deadcode analysis failure.
On Mac the cgroups package cannot be loaded, and this will cause `make
check` to fail in the go114-macos job.

Move the cgroups checks to `require_linux.go` to avoid this failure.
