+++ internal/pkg/test/tool/require/require_linux.go
@@ -14,6 +14,7 @@ import (
 	"syscall"
 	"testing"
 
+	"github.com/containerd/cgroups"
 	"github.com/sylabs/singularity/internal/pkg/buildcfg"
 	"github.com/sylabs/singularity/pkg/network"
 )
@@ -110,4 +111,29 @@ func Network(t *testing.T) {
 	if !supportNetwork {
 		t.Skipf("Network (bridge) not supported")
 	}
+}
+
+// Cgroups checks that cgroups is enabled, if not the
+// current test is skipped with a message.
+func Cgroups(t *testing.T) {
+	_, err := cgroups.V1()
+	if err != nil {
+		t.Skipf("cgroups disabled")
+	}
+}
+
+// CgroupsFreezer checks that cgroup freezer subsystem is
+// available, if not the current test is skipped with a
+// message
+func CgroupsFreezer(t *testing.T) {
+	subSys, err := cgroups.V1()
+	if err != nil {
+		t.Skipf("cgroups disabled")
+	}
+	for _, s := range subSys {
+		if s.Name() == "freezer" {
+			return
+		}
+	}
+	t.Skipf("no cgroups freezer subsystem available")
 }
